Move computer keys to structure and drop all old code
