+++ pdbufr/bufr_structure.py
@@ -1,4 +1,5 @@
 import collections
+import datetime
 import typing as T
 
 import attr
@@ -110,6 +111,46 @@ def filter_keys_cached(
     return cache[filtered_message_uid]
 
 
+def datetime_from_bufr(observation, prefix, datetime_keys):
+    # type: (T.Dict[str, T.Any], str, T.List[str]) -> pd.Timestamp
+    hours = observation.get(prefix + datetime_keys[3], 0)
+    minutes = observation.get(prefix + datetime_keys[4], 0)
+    seconds = observation.get(prefix + datetime_keys[5], 0.0)
+    microseconds = int(seconds * 1_000_000) % 1_000_000
+    datetime_list = [observation[prefix + k] for k in datetime_keys[:3]]
+    datetime_list += [hours, minutes, int(seconds)]
+    return datetime.datetime(*datetime_list, microsecond=microseconds)
+
+
+def wmo_station_id_from_bufr(observation, prefix, keys):
+    # type: (T.Dict[str, T.Any], str, T.List[str]) -> int
+    block_number = int(observation[prefix + keys[0]])
+    station_number = int(observation[prefix + keys[1]])
+    return block_number * 1000 + station_number
+
+
+COMPUTED_KEYS = [
+    (
+        ["year", "month", "day", "hour", "minute", "second"],
+        "data_datetime",
+        datetime_from_bufr,
+    ),
+    (
+        [
+            "typicalYear",
+            "typicalMonth",
+            "typicalDay",
+            "typicalHour",
+            "typicalMinute",
+            "typicalSecond",
+        ],
+        "typical_datetime",
+        datetime_from_bufr,
+    ),
+    (["blockNumber", "stationNumber"], "WMO_station_id", wmo_station_id_from_bufr),
+]
+
+
 def extract_observations(
     message: T.Mapping[str, T.Any],
     filtered_keys: T.List[BufrKey],
@@ -180,7 +221,7 @@ def add_computed_keys(
     observation: T.Dict[str, T.Any], included_keys: T.Container[str]
 ) -> T.Dict[str, T.Any]:
     augmented_observation = observation.copy()
-    for keys, computed_key, getter in bufr_read.COMPUTED_KEYS:
+    for keys, computed_key, getter in COMPUTED_KEYS:
         if computed_key not in included_keys:
             continue
         computed_value = getter(observation, "", keys)
@@ -188,7 +229,7 @@ def add_computed_keys(
     return augmented_observation
 
 
-def filter_stream(
+def stream_bufr(
     bufr_file: T.Iterable[T.MutableMapping[str, T.Any]],
     columns: T.Iterable[str],
     filters: T.Mapping[str, T.Any] = {},
@@ -219,7 +260,7 @@ def filter_stream(
     compiled_filters = bufr_filters.compile_filters(filters)
     included_keys = set(compiled_filters)
     included_keys |= set(columns)
-    for keys, computed_key, _ in bufr_read.COMPUTED_KEYS:
+    for keys, computed_key, _ in COMPUTED_KEYS:
         if computed_key in included_keys:
             included_keys |= set(keys)
